% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_digit_small.R
\name{predict_digit_small}
\alias{predict_digit_small}
\title{Predicts the digit based on a gray-scale image}
\usage{
predict_digit_small(image)
}
\arguments{
\item{image}{A 784 (28x28) length vector with gray-scale values between 0-255.}
}
\value{
The predicted digit
}
\description{
Predicts the digit based on a gray-scale image
}
